Add files via upload
Gone are the days when you had to manually browse and spend hours searching for different items on various websites. Today that has become much easier with the development of recommender systems that brings everything you need right before you as soon as you visit any website. You must be wondering how this is done.

 

Recommender systems are one of the most frequently implemented and in-demand applications of machine learning in the industry. Almost every product company that sells a consumer-based product requires a recommender system to be deployed on its platform.  Revenue generation of any firm depends on the quality of user experience it offers to its customers, and recommender systems play a significant role in improving this. It covers almost every aspect of machine learning right from natural language processing to deep learning. 

 

It is a type of information filtering system that figures out user preferences based on the choices they make. Customers find personalization appealing, enhancing customer retention and increased sales leading to higher revenue generation and profits.

 

Recommendation systems provide personalized user experiences in almost all domains, such as online shopping, news websites, gaming, and social media platforms.

 

In this module
You will get to learn about:

Recommendation Systems
Different Types of Recommendation Systems which include content-based models and collaborative filtering models and use them to build a hybrid model to generate recommendations for a given user or an item. Here is a mindmap of the types of recommendation systems that you will be learning in this module:
Recommendation Systems
Recommendation Systems
Evaluation Methods
Issues Faced in Recommendation Systems
Building a Movie Recommender System in Python
 

The final session of the module defines the problem statement and the deliverables of the project.

 

Through this project, you will be building a news recommender system that recommends different news articles to users based on their reading preferences. The most popular example of a news recommendation system is Google News, which personalizes your news feed based on your interests and favourite topics. You will be able to deploy a real-world application, and, in the process, learn many useful skills required to develop and deploy ML products.